# Add taylor to lookup