This is the same xsimd lookup with a Taylor expansion.

We get now a precision of 1.e-6.