@wvbbreu 2025-08-13
For reference, see the performance results below.

**Old implementation**

```
benchmark_sinf<Backend>/100000           0.054 ms        0.053 ms        13251 init_ms=0 items_per_second=1.87013G/s
benchmark_sinf<Backend>/1000000          0.543 ms        0.542 ms         1295 init_ms=0 items_per_second=1.84438G/s
benchmark_sinf<Backend>/10000000          5.56 ms         5.55 ms          127 init_ms=0 items_per_second=1.80286G/s
benchmark_cosf<Backend>/100000           0.055 ms        0.055 ms        12735 init_ms=0 items_per_second=1.82012G/s
benchmark_cosf<Backend>/1000000          0.559 ms        0.558 ms         1255 init_ms=0 items_per_second=1.79338G/s
benchmark_cosf<Backend>/10000000          5.84 ms         5.82 ms          122 init_ms=0 items_per_second=1.71809G/s
benchmark_sincosf<Backend>/100000        0.111 ms        0.111 ms         6310 init_ms=0 items_per_second=903.028M/s
benchmark_sincosf<Backend>/1000000        1.09 ms         1.09 ms          636 init_ms=0 items_per_second=918.117M/s
benchmark_sincosf<Backend>/10000000       11.5 ms         11.5 ms           61 init_ms=0 items_per_second=871.938M/s
benchmark_sinf<Backend>/100000           0.054 ms        0.054 ms        13049 init_ms=0 items_per_second=1.86367G/s
benchmark_sinf<Backend>/1000000          0.543 ms        0.541 ms         1294 init_ms=0 items_per_second=1.84714G/s
benchmark_sinf<Backend>/10000000          5.53 ms         5.52 ms          127 init_ms=0 items_per_second=1.81247G/s
benchmark_cosf<Backend>/100000           0.055 ms        0.055 ms        12699 init_ms=0 items_per_second=1.81559G/s
benchmark_cosf<Backend>/1000000          0.557 ms        0.556 ms         1256 init_ms=0 items_per_second=1.79798G/s
benchmark_cosf<Backend>/10000000          5.78 ms         5.76 ms          121 init_ms=0 items_per_second=1.73482G/s
benchmark_sincosf<Backend>/100000        0.109 ms        0.109 ms         6423 init_ms=0 items_per_second=918.204M/s
benchmark_sincosf<Backend>/1000000        1.09 ms         1.09 ms          641 init_ms=0 items_per_second=915.628M/s
benchmark_sincosf<Backend>/10000000       11.4 ms         11.4 ms           61 init_ms=0 items_per_second=875.501M/s
```

**New implementation**

```
benchmark_sinf<Backend>/100000           0.108 ms        0.107 ms         6541 init_ms=0 items_per_second=931.429M/s
benchmark_sinf<Backend>/1000000           1.08 ms         1.08 ms          648 init_ms=0 items_per_second=926.274M/s
benchmark_sinf<Backend>/10000000          10.8 ms         10.8 ms           65 init_ms=0 items_per_second=924.239M/s
benchmark_cosf<Backend>/100000           0.108 ms        0.108 ms         6473 init_ms=0 items_per_second=925.144M/s
benchmark_cosf<Backend>/1000000           1.09 ms         1.09 ms          643 init_ms=0 items_per_second=920.43M/s
benchmark_cosf<Backend>/10000000          10.9 ms         10.9 ms           64 init_ms=0 items_per_second=918.406M/s
benchmark_sincosf<Backend>/100000        0.119 ms        0.118 ms         5913 init_ms=0 items_per_second=844.396M/s
benchmark_sincosf<Backend>/1000000        1.19 ms         1.19 ms          589 init_ms=0 items_per_second=840.596M/s
benchmark_sincosf<Backend>/10000000       11.9 ms         11.8 ms           59 init_ms=0 items_per_second=844.746M/s
benchmark_sinf<Backend>/100000           0.111 ms        0.110 ms         6346 init_ms=0 items_per_second=906.621M/s
benchmark_sinf<Backend>/1000000           1.11 ms         1.11 ms          631 init_ms=0 items_per_second=901.669M/s
benchmark_sinf<Backend>/10000000          11.3 ms         11.2 ms           62 init_ms=0 items_per_second=890.302M/s
benchmark_cosf<Backend>/100000           0.112 ms        0.111 ms         6288 init_ms=0 items_per_second=898.318M/s
benchmark_cosf<Backend>/1000000           1.12 ms         1.12 ms          625 init_ms=0 items_per_second=891.996M/s
benchmark_cosf<Backend>/10000000          10.9 ms         10.9 ms           64 init_ms=0 items_per_second=918.038M/s
benchmark_sincosf<Backend>/100000        0.121 ms        0.120 ms         5814 init_ms=0 items_per_second=830.322M/s
benchmark_sincosf<Backend>/1000000        1.21 ms         1.21 ms          579 init_ms=0 items_per_second=826.209M/s
benchmark_sincosf<Backend>/10000000       12.3 ms         12.2 ms           57 init_ms=0 items_per_second=816.368M/s
```